Feature/class infrastructure (#7)
Created basic classes for servers, items, files, etc.
This is meant to generalize the flow of the program and standardize the API with different file servers.
While still lacking (mostly in the promos area), the fundamentals are built. Parsing promos is still problematic, as they are formatted in different ways across servers; meaning, a better parsing system needs to be built to handle parsing of the fetched files in a more general way. Albeit, this progression needs to happen after setting up the database - mainly defining it's schema. This will help to see how we need to define the data and therefore how to parse it.